separate out states
States are now separate files under the server so that they aren't
exposed to downstream consumers.
