+++ crates/apollo-mcp-server/src/server/states/operations_configured.rs
@@ -1,0 +1,54 @@
+use apollo_compiler::{Schema, validation::Valid};
+use reqwest::header::HeaderMap;
+use tracing::debug;
+
+use crate::{
+    custom_scalar_map::CustomScalarMap,
+    errors::ServerError,
+    operations::{MutationMode, RawOperation},
+    server::{Transport, states::Starting},
+};
+
+pub(super) struct OperationsConfigured {
+    pub(super) transport: Transport,
+    pub(super) operations: Vec<RawOperation>,
+    pub(super) endpoint: String,
+    pub(super) headers: HeaderMap,
+    pub(super) introspection: bool,
+    pub(super) explorer_graph_ref: Option<String>,
+    pub(super) custom_scalar_map: Option<CustomScalarMap>,
+    pub(super) mutation_mode: MutationMode,
+    pub(super) disable_type_description: bool,
+    pub(super) disable_schema_description: bool,
+}
+
+impl OperationsConfigured {
+    pub(super) async fn set_schema(self, schema: Valid<Schema>) -> Result<Starting, ServerError> {
+        debug!("Received schema:\n{}", schema);
+        Ok(Starting {
+            transport: self.transport,
+            schema,
+            operations: self.operations,
+            endpoint: self.endpoint,
+            headers: self.headers,
+            introspection: self.introspection,
+            explorer_graph_ref: self.explorer_graph_ref,
+            custom_scalar_map: self.custom_scalar_map,
+            mutation_mode: self.mutation_mode,
+            disable_type_description: self.disable_type_description,
+            disable_schema_description: self.disable_schema_description,
+        })
+    }
+
+    pub(super) async fn set_operations(
+        self,
+        operations: Vec<RawOperation>,
+    ) -> Result<OperationsConfigured, ServerError> {
+        debug!(
+            "Received {} operations:\n{}",
+            operations.len(),
+            serde_json::to_string_pretty(&operations)?
+        );
+        Ok(OperationsConfigured { operations, ..self })
+    }
+}
